pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "flink-cdc-gateway"
dynamic = ["version"]
description = "Change Data Capture Gateway for Apache Flink"
readme = "README.md"
authors = [
    {name = "Moshe Eliya", email = "mosiko1234@gmail.com"}
]
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Database",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "flask>=2.0.0",
    "flask-restful>=0.3.9",
    "requests>=2.25.1",
    "pyyaml>=6.0",
    "jsonschema>=4.0.0",
    "kafka-python>=2.0.2",
    "prometheus-client>=0.11.0",
    "python-dotenv>=0.19.0",
    "gunicorn>=20.1.0",
    "psutil>=5.9.0",
]

[project.urls]
Homepage = "https://github.com/mosiko1234/flink-cdc-gateway"
"Bug Tracker" = "https://github.com/mosiko1234/flink-cdc-gateway/issues"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "flake8>=4.0.0",
    "black>=22.3.0",
    "isort>=5.10.0",
    "mypy>=0.950",
]

[project.scripts]
cdc-gateway = "cdc_gateway.app:main"
cdc-gateway-admin = "cdc_gateway.admin:main"
cdc-gateway-monitor = "cdc_gateway.monitor:main"

[tool.setuptools]
packages = ["cdc_gateway"]

[tool.setuptools.dynamic]
version = {attr = "cdc_gateway.__version__"}

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[tool.black]
line-length = 100
target-version = ["py38"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false